Started test cases for auth
[Ticket: #176]
